+++ lib/api.dart
@@ -6,10 +6,12 @@ import 'dart:typed_data';
 import 'package:convert/convert.dart';
 import 'package:pinenacl/crypto/tweetnacl.dart';
 
+export 'package:pinenacl/crypto/hmac_sha512.dart';
 export 'package:pinenacl/crypto/blake2b.dart';
 export 'package:pinenacl/crypto/tweetnacl.dart';
 export 'package:pinenacl/public.dart';
 export 'package:pinenacl/signing.dart';
+export 'package:pinenacl/hashing.dart';
 export 'package:pinenacl/utils.dart';
 
 part 'src/api/asymmetric.dart';