+++ lib/crypto/hmac_sha512.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 import 'package:convert/convert.dart';
 
-import 'tweetnacl.dart';
+import 'package:pinenacl/api.dart';
 
 /// https://csrc.nist.gov/csrc/media/publications/fips/198/1/final/documents/fips-198-1_final.pdf
 /// HMAC-SHA-512 implementation