+++ CHANGELOG.md
@@ -39,3 +39,4 @@ and `crypto_auth_verify` functions of the `NaCl` library (does not exists in Twe
 - Added HMAC-SHA-512 unit tests.
 - Added some `TweetNaCl`'s tests.
 - Cleaned some code.
+- Fixed exports.